fix(core): align iOS dummy Link::num_peers return type with host (usize)
The clippy cleanup (845f92c) dropped `link.num_peers() as usize` as a "redundant"
cast — true on the host build (real BasicLink::num_peers -> usize), but the iOS
cfg dummy returned u32, so the cast was load-bearing for the iOS cross-compile
(4 errors: `peers` stored into AtomicUsize, compared against a usize, and passed
to LinkPeersChanged { count: usize }). The host `cargo clippy`/`cargo test` gate
is blind to #[cfg(target_os = "ios")] paths.

Make the dummy return usize to match the host signature; the cast-free call site
is correct on both targets. Verified: `cargo check --target aarch64-apple-ios`
green, host `cargo clippy --all-targets -- -D warnings` green.

+++ engine/crates/core/src/engine.rs
@@ -37,7 +37,7 @@ impl Link {
     pub fn clock(&self) -> DummyClock {
         DummyClock
     }
-    pub fn num_peers(&self) -> u32 {
+    pub fn num_peers(&self) -> usize {
         0
     }
 }
